py-graph/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
ansible==9.10.0
ansible-core==2.16.11
anyio==3.5.0
apache-libcloud==3.2.0
appdirs==1.4.4
argcomplete==1.8.1
asitop==0.0.24
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==21.2.0
Babel==2.8.0
backoff==2.2.1
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==21.10b0
blessed==1.20.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.2
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==2.0.6
chrome-gnome-shell==0.0.0
click==8.0.3
click-default-group==1.2.4
colorama==0.4.4
colorcet==3.0.1
comm==0.2.1
command-not-found==0.3
commonmark==0.9.1
contourpy==1.2.0
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dashing==0.1.0
dataclasses-json==0.6.2
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
diskcache==5.6.3
distro==1.7.0
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
editor==1.6.5
exceptiongroup==1.2.0
execnet==1.9.0
executing==2.0.1
filelock==3.13.1
flake8==4.0.1
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2023.12.2
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.41
gpt4all==2.0.2
greenlet==3.0.1
h11==0.13.0
hidpidaemon==18.4.6
html2image==2.0.4.3
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.20.2
idna==3.3
imageio==2.33.1
importlib-metadata==7.0.1
iniconfig==1.1.1
inquirer==3.2.1
ipykernel==6.28.0
ipython==8.20.0
isort==5.6.4
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
jupyter_client==8.6.0
jupyter_core==5.7.1
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.0.340
langchain-community==0.0.13
langchain-core==0.1.11
langsmith==0.0.66
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
litellm==1.17.0
livereload==2.6.3
llama_cpp_python==0.2.28
llm-gpt4all==0.2
llm-llama-cpp==0.3
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mkdocs==1.1.2
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
multidict==6.0.4
mutagen==1.45.1
mypy==0.942
mypy-extensions==0.4.3
nala==0.11.1
nest-asyncio==1.5.8
netaddr==0.8.0
netifaces==0.11.0
nltk==3.8.1
nose2==0.9.2
notify2==0.3
ntlm-auth==1.4.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
ollama==0.3.1
open-interpreter==0.2.0
openai==0.28.1
openshot-qt==2.5.1
outcome==1.3.0.post0
packaging==23.2
param==2.0.1
paramiko==2.9.3
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==10.1.0
pip-autoremove==0.10.0
platformdirs==2.5.1
pluggy==0.13.0
pop-transition==1.1.2
posthog==3.3.1
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
PyAutoGUI==0.9.54
pycairo==1.20.1
pycodestyle==2.8.0
pycryptodomex==3.11.0
pyct==0.5.0
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydbus==0.6.0
pyflakes==2.4.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.30
PySocks==1.7.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-sugar==0.9.4
pytest-xdist==2.5.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-ulid==2.2.0
python-xlib==0.29
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.1
pywinrm==0.3.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
ranger-fm==1.9.3
readchar==4.0.5
regex==2023.12.25
repolib==2.2.1
repoman==1.4.0
reportlab==3.6.8
requests==2.31.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
resolvelib==0.8.1
rfc3986==1.5.0
rich==13.7.0
runs==1.2.0
scour==0.38.2
SecretStorage==3.3.1
selenium==4.16.0
selinux==3.3
sessioninstaller==0.0.0
shellingham==1.4.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.10.4
speedtest-cli==2.1.3
SQLAlchemy==2.0.23
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.36
ssh-import-id==5.11
stack-data==0.6.3
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==1.1.0
texttable==1.6.4
tiktoken==0.4.0
tokenizers==0.15.0
tokentrim==0.1.13
toml==0.10.2
tomli==1.2.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trash-cli==0.17.1.14
trio==0.23.2
trio-websocket==0.11.1
typed-ast==1.4.3
typer==0.4.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
urllib3==1.26.5
urwid==2.1.2
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==9.1
wget==3.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
xmltodict==0.12.0
xmod==1.8.1
yarl==1.9.3
yt-dlp==2022.4.8
zipp==1.0.0
